修复订单删除问题

+++ yshop-shop/src/main/java/co/yixiang/modules/order/rest/StoreOrderController.java
@@ -268,7 +268,7 @@ public class StoreOrderController {
     @ApiOperation(value = "删除")
     @DeleteMapping(value = "/yxStoreOrder/{id}")
     @PreAuthorize("hasAnyRole('admin','YXSTOREORDER_ALL','YXSTOREORDER_DELETE')")
-    public ResponseEntity delete(@PathVariable Integer id) {
+    public ResponseEntity delete(@PathVariable Long id) {
         yxStoreOrderService.removeById(id);
         return new ResponseEntity(HttpStatus.OK);
     }
